backend/requirements.txt
annotated-types==0.6.0
adal==1.2.7
anyio==3.7.1
bcrypt==4.1.3
click==8.1.7
dnspython==2.6.1
ecdsa==0.18.0
email-validator==1.3.1
exceptiongroup==1.1.3
fastapi==0.110.0
greenlet==3.0.1
h11==0.14.0
idna==3.4
itsdangerous==2.2.0
Jinja2==3.0.3
MarkupSafe==2.1.5
passlib==1.7.4
pyasn1==0.5.1
pydantic==1.10.13
pydantic_core==2.10.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.9
requests==2.32.3
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.50
SQLAlchemy-Utils==0.41.1
sqlalchemy2-stubs==0.0.2a36
sqlmodel==0.0.11
starlette==0.36.3
starlette-admin==0.13.2
typing_extensions==4.8.0
uvicorn==0.24.0.post1